+++ src/preprocessing/max_abs_scaler/mod.rs
@@ -5,9 +5,9 @@ use crate::artifact::{ArtifactError, MAX_ABS_SCALER_ARTIFACT_KIND};
 use crate::data::MatrixView;
 
 use super::scaling::{
-    ScalerHeader, ScalerParameters, decode_scaler_artifact, encode_scaler_artifact,
-    inverse_transform_allocating, substituted_divisor, transform_preflighted,
-    validate_inverse_request, validate_transform_request,
+    BASE_PAYLOAD_VERSION, ScalerHeader, ScalerParameters, decode_scaler_artifact,
+    encode_scaler_artifact, inverse_transform_allocating, substituted_divisor,
+    transform_preflighted, validate_inverse_request, validate_transform_request,
 };
 
 /// Parameters for [`MaxAbsScaler`].
@@ -138,6 +138,7 @@ impl MaxAbsScaler {
             transformed_schema,
             self.n_features_in,
             ScalerParameters {
+                version: BASE_PAYLOAD_VERSION,
                 flags: &[],
                 reals: &[],
             },
@@ -161,6 +162,7 @@ impl MaxAbsScaler {
             MAX_ABS_SCALER_ARTIFACT_KIND,
             input_schema,
             transformed_schema,
+            BASE_PAYLOAD_VERSION,
             0,
             0,
         )?;